requirements.txt
Flask==0.11.1
requests>=2.8.1
gcloud==0.15.0
gunicorn==19.6.0
oauth2client==2.2.0
six==1.10.0
werkzeug
GoogleAppEngineCloudStorageClient
clarifai
elasticsearch>=2.0.0
